+++ bd/src/main/java/org/example/test24/bd/BaseData1ClassMsSql.java
@@ -226,7 +226,7 @@ public class BaseData1ClassMsSql extends BaseData1Class {
                     "UPDATE Table_users SET  \"password\" = ? WHERE \"id\" = ?"
             );
             statement.setString(1, pass);
-            statement.setInt(2, userClass.id_user);
+            statement.setLong(2, userClass.id_user);
             statement.executeUpdate();
             statement.close();
         } catch (java.lang.Throwable ex) {
