+++ bd/src/main/java/org/example/test24/bd/User.java
@@ -5,15 +5,15 @@ import java.util.Date;
 public class User {
     final public static int RANG_USERS = 0;
     final public static int RANG_PUSHERS = 1;
-    public int id_user;
+    public long id_user;
     public Date date;
-    public int id_loggerUser;
+    public long id_loggerUser;
     public String name;
     public String password;
     public int rang;
     public Date date_unreg;
 
-    public User(int id_user, Date date, int id_loggerUser, String name, String password, int rang, Date date_unreg) {
+    public User(long id_user, Date date, long id_loggerUser, String name, String password, int rang, Date date_unreg) {
         this.id_user = id_user;
         this.date = date;
         this.id_loggerUser = id_loggerUser;
